Ajout des EventBus pour les taches avec les notifications toast + Modification de la logique routes Task

+++ client/src/components/TaskCard.jsx
@@ -2,15 +2,19 @@ import Button from 'react-bootstrap/Button';
 import { useState } from 'react';
 import Card from 'react-bootstrap/Card';
 import CloseButton from 'react-bootstrap/CloseButton';
+import { sendMessage } from '../service/webSocketService';
+import Task from '../../../server/schemas/taskSchema';
 
 
-function TaskCard({ task, fetchTasks }) {
+function TaskCard({ task, fetchTasks, user }) {
   const [isModify, setIsModify] = useState(false);
   const [isCompleted, setIsCompleted] = useState(task.completed);
   const [title, setTtitle] = useState(task.title || "");
   const [description, setDescription] = useState(task.description || "");
   const created_at = task.created_at;
   const category_id = task.category_id;
+  const taskId = task._id;
+
 
   const handleCompleted = () => {
     setIsCompleted(!isCompleted);
@@ -24,23 +28,28 @@ function TaskCard({ task, fetchTasks }) {
 
   const handleDelete = async () => {
     try {
-      const response = await fetch(`http://localhost:3001/api/tasks/${task._id}`, {
-        method: 'DELETE',
-        credentials: "include"
-      });
 
-      const data = await response.json();
+      if (!user) {
 
-      if (data.source === "db") {
-        if (!response.ok) {
-          throw new Error(data.message || `Erreur ${response.status}`);
-        }
-      } else if (data.source === "Guest") {
         const localTasks = JSON.parse(localStorage.getItem("defaultTasks"));
-        const newTasks = localTasks.filter(task => task._id != data.id);
+        const newTasks = localTasks.filter(task => task._id !== taskId);
         localStorage.setItem("defaultTasks", JSON.stringify(newTasks));
       }
 
+      if (user) {
+        const response = await fetch(`http://localhost:3001/api/tasks/${task._id}`, {
+          method: 'DELETE',
+          credentials: "include"
+        });
+
+        const data = await response.json();
+        console.log(data)
+        sendMessage({ type: "DELETED_TASK", payload: data.deletedTask });
+        
+        if (!response.ok) {
+          throw new Error(data.message || `Erreur ${response.status}`);
+        }
+      }
     } catch (error) {
       console.error(error)
     } finally {
@@ -53,26 +62,29 @@ function TaskCard({ task, fetchTasks }) {
 
 
     try {
-      const response = await fetch(`http://localhost:3001/api/tasks/${task._id}`, {
-        method: "PUT",
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ title, description, created_at, category_id }),
-        credentials: "include"
-      })
+      if (!user) {
+        const localTask = JSON.parse(localStorage.getItem("defaultTasks"));
+        const currentTask = localTask.filter(task => task._id === taskId);
+        const newTask = {...currentTask[0], title:title, description:description, updated_at: new Date().toString()}
+        const newLocalTask = [...localTask.filter(task => task._id !== newTask._id), newTask]
+        localStorage.setItem("defaultTasks", JSON.stringify(newLocalTask));
+      }
 
-      const data = await response.json();
+      if (user) {
+        const response = await fetch(`http://localhost:3001/api/tasks/${task._id}`, {
+          method: "PUT",
+          headers: { 'Content-Type': 'application/json' },
+          body: JSON.stringify({ title, description, created_at, category_id }),
+          credentials: "include"
+        })
 
-      if (data.source === "db") {
-        console.log("la response est ", data)
+        const data = await response.json();
+
+        sendMessage({ type: "UPDATE_TASK", payload: data.updatedTask });
 
         if (!response.ok) {
           throw new Error("Erreur")
         }
-      } else if (data.source === "Guest") {
-        const localTask = JSON.parse(localStorage.getItem("defaultTasks"));
-        const newTask = { ...data.taskModify, completed: isCompleted };
-        const newLocalTask = [...localTask.filter(task => task._id !== newTask._id), newTask]
-        localStorage.setItem("defaultTasks", JSON.stringify(newLocalTask));
       }
     } catch (error) {
       console.error(error)
@@ -103,6 +115,8 @@ function TaskCard({ task, fetchTasks }) {
       </div>
     )
   }
+
+
 
   return (
     <div className="border border-warning rounded-2 ps-2 w-75 mx-auto my-4">
